Split HDF5 tool into multiple logical subfiles for ease of maintaining
